+++ src/components/view/ContactIntake/newContactQuestions.form.js
@@ -3,6 +3,8 @@ import React, { Component } from 'react'
 import { Card, CardTitle } from 'material-ui/Card';
 import DatePicker from 'material-ui/DatePicker';
 
+import {radioButtonGenderList, enthnicityOptionsList, countryOfBirthList } from '../../../utils/fieldValueLists';
+
 
 class NewContactQuestionsForm extends Component {
 
@@ -19,32 +21,6 @@ class NewContactQuestionsForm extends Component {
     render() {
 
         const { palette, updateIntakeFormField } = this.props;
-
-        const radioButtonGenderList = [
-            {label: 'Male', value: 'male'},
-            {label: 'Female', value: 'female'},
-            {label: 'MtF', value: 'mtf'},
-            {label: 'FtM', value: 'ftm'},
-            {label: 'Refused', value: 'refused'},
-        ];
-
-        const enthnicityOptionsList = [
-            {primaryText: 'American Indian', value: 'AI'},
-            {primaryText: 'Asian', value: 'Asian'},
-            {primaryText: 'Black/African American', value: 'BAA'},
-            {primaryText: 'Hawaiin/Pacific Islander', value: 'HPI'},
-            {primaryText: 'White', value: 'White'},
-            // TODO: user needs to be able to click this and enter in their own custom entry
-            {primaryText: 'Other', value: 'Other'},
-            {primaryText: 'Prefer not to say', value: 'Refuse'},
-        ];
-
-        const countryOfBirthList = [
-            {label: 'US', value: 'US'},
-            {label: 'Puerto Rico', value: 'PR'},
-            // TODO: user needs to be able to click this and enter in their own custom entry
-            {label: 'Other', value: 'Other'},
-        ];
 
         const options = {
             defaultValue: 0,
