+++ src/components/view/ContactIntake/periodicQuestions.form.js
@@ -5,6 +5,7 @@ import { FieldWithManualOption } from '../common/FieldWithManualOption';
 import { Card, CardTitle } from 'material-ui/Card';
 import TextField from 'material-ui/TextField';
 
+import { housingRadioOptions, hivRadioOptions, hepCOptions, insurerOptionsList, primaryDrugOptionsList, otherDrugOptionsList } from '../../../utils/fieldValueLists';
 
 class PeriodicIntakeForm extends Component {
 
@@ -24,55 +25,6 @@ class PeriodicIntakeForm extends Component {
     render() {
 
         const { updateIntakeFormField } = this.props;
-
-        const housingRadioOptions = [
-            { name: 'housed', label: 'Housed', value: 'housed' },
-            { name: 'homeless', label: 'Homeless', value: 'homeless' },
-            { name: 'unstable', label: 'Unstable Housed', value: 'unstable' },
-        ];
-
-        const hivRadioOptions = [
-            { name: 'positive', label: 'Positive', value: 'positive' },
-            { name: 'negative', label: 'Negative', value: 'negative' },
-            { name: 'neverTested', label: 'Never Tested', value: 'neverTested' },
-        ]
-
-        const hepCOptions = [
-            { name: 'positive', label: 'Positive', value: 'positive' },
-            { name: 'negative', label: 'Negative', value: 'negative' },
-            { name: 'neverTested', label: 'Never Tested', value: 'neverTested' },
-        ];
-
-        const insurerOptionsList = [
-            {primaryText: 'Mass Health', value: 'MH'},
-            {primaryText: 'Neighborhood Health Program', value: 'NHP'},
-            {primaryText: 'Blue Cross Blue Shield', value: 'BCBS'},
-            {primaryText: 'Tufts', value: 'TF'},
-            // TODO: user needs to be able to click this and enter in their own custom entry
-            {primaryText: 'Other', value: 'other'},
-        ];
-
-        const primaryDrugOptionsList = [
-            {primaryText: '-none-', value: null},
-            {primaryText: 'Heroin', value: 'heroin'},
-            {primaryText: 'Methamphetamine', value: 'meth'},
-            {primaryText: 'Cocaine', value: 'cocaine'},
-            // TODO: user needs to be able to click this and enter in their own custom entry
-            {primaryText: 'Other', value: 'other'},
-        ];
-
-        const otherDrugOptionsList = [
-            {primaryText: '-none-', value: null},
-            {primaryText: 'Heroin', value: 'heroin'},
-            {primaryText: 'Methamphetamine', value: 'meth'},
-            {primaryText: 'Cocaine', value: 'cocaine'},
-            {primaryText: 'Crack', value: 'crack'},
-            {primaryText: 'Benzos', value: 'benzos'},
-            {primaryText: 'Suboxone (not RX)', value: 'suboxone'},
-            {primaryText: 'Methadone (not RX)', value: 'methadone'},
-            // TODO: user needs to be able to click this and enter in their own custom entry
-            {primaryText: 'Other', value: 'other'},
-        ]
 
         // rest of the stuff for this form
         return (
