+++ src/sagas/reports.js
@@ -1,30 +1,12 @@
 import { takeEvery } from 'redux-saga/effects';
 
 import {
-    RANGE_ALL_TIME,
     FETCH_REPORTS_DATA,
-    RANGE_CURRENT_WEEK,
-    RANGE_PREVIOUS_WEEK,
-    RANGE_CURRENT_YEAR,
-    RANGE_PREVIOUS_YEAR,
 } from '../constants';
 
-import Moment from 'moment';
-import { extendMoment } from 'moment-range';
+import { getDateBoundsFromRangeKey } from '../utils/dateRangeUtils';
 
 import { updateReportsData } from '../actions';
-
-// memoize moment assignment
-const getExtendedMoment = (() => {
-
-    let cachedMoment = null;
-    
-    return getCachedMoment;
-
-    function getCachedMoment() {
-        return cachedMoment || ( cachedMoment = extendMoment(Moment) );
-    }
-})();
 
 function* fetchReportsData({ dateRange }) {
 
@@ -180,46 +162,4 @@ function getContactBreakdownData ( contacts ) {
 
         return agg;
     }, {});
-}
-
-function getDateBoundsFromRangeKey({ rangeKey }) {
-
-    const moment = getExtendedMoment();
-    const now = new Date();
-    let min, max;
-
-    // note isoWeek starts on Monday
-
-    switch ( rangeKey ) {
-        case RANGE_ALL_TIME:
-            min = moment('2017');
-            max = moment(now).add(1, 'day');
-            break;
-
-        case RANGE_PREVIOUS_WEEK:
-            min = moment().startOf('isoWeek').subtract(1, 'week');
-            max = moment().startOf('isoWeek').subtract(1, 'day');
-            break;
-
-        case RANGE_CURRENT_YEAR:
-            min = moment().startOf('year');
-            max = moment(now).add(1, 'day');
-            break;
-
-        case RANGE_PREVIOUS_YEAR:
-            min = moment().startOf('year').subtract(1, 'year');
-            max = moment().startOf('year').subtract(1, 'day');
-            break;
-
-        case RANGE_CURRENT_WEEK:
-        default:
-            min = moment().startOf('isoWeek');
-            max = moment(now).add(1, 'day');
-            break;
-    }
-
-    return {
-        min: min.toDate(),
-        max: max.toDate(),
-    };
 }
